backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
qdrant-client[fastembed]
sentence-transformers
langdetect
sqlalchemy
python-multipart
httpx
openai
google-genai
